pyproject.toml
[tool.poetry]
name = "telephant-server"
version = "0.1.0"
description = "Telephant Server + API"
authors = ["Tomas Hlavacek <tmshlvck@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
#packages = [{include = "telephant_server"},]# {include = "webcrud"}]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.24.0.post1"
pyyaml = "^6.0.1"
fastapi = "^0.104.1"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
#sqlalchemy = {version = "*", extras = ["sqlite3"]}
sqlalchemy = "^2.0.23"
authlib = "^1.2.1"
httpx = "^0.25.2"
itsdangerous = "^2.1.2"
pyjwt = "^2.8.0"
sqlmodel = "^0.0.14"
passlib = "^1.7.4"
requests = "^2.31.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
telephant_server = "telephant_server:main"
